Code clean up
Remove duplicate from all the resourcesync xml classes
