+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/dataset/DataSetFactory.java
@@ -152,6 +152,7 @@ public class DataSetFactory implements DataFetcherFactoryFactory, SchemaStoreFac
         dataStoreFactory.createDataSourceStore(importManager, userId, dataSetId)
       );
 
+      this.configuration.getResourceSync().addDataSet(userId, dataSetId);
       return dataSet;
     } catch (AuthorizationCreationException | IOException e) {
       throw new DataStoreCreationException(e);
