+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/datastores/resourcesync/FileSystemResourceList.java
@@ -4,18 +4,8 @@ import nl.knaw.huygens.timbuctoo.v5.filestorage.dto.CachedFile;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
-import org.xml.sax.SAXException;
 
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
 import java.io.File;
-import java.io.IOException;
 
 class FileSystemResourceList implements ResourceList {
   private final File resourceList;
@@ -27,58 +17,10 @@ class FileSystemResourceList implements ResourceList {
   }
 
   @Override
-  public void addFile(CachedFile fileToAdd) {
-    DocumentBuilderFactory docFactory = DocumentBuilderFactory.newInstance();
-    DocumentBuilder docBuilder = null;
-    try {
-      docBuilder = docFactory.newDocumentBuilder();
-    } catch (ParserConfigurationException e) {
-      e.printStackTrace();
-    }
-    Node root;
-    Document doc;
-    if (!resourceList.exists()) {
-      doc = docBuilder.newDocument();
-      root = createRootNode(doc);
-      doc.appendChild(root);
-    } else {
-      try {
-        doc = docBuilder.parse(resourceList);
-
-      } catch (SAXException | IOException e) {
-        throw new RuntimeException(e);
-      }
-      root = doc.getFirstChild();
-    }
-
-    updateMetaData(root, doc);
-
-    root.appendChild(createUrlNode(doc, fileToAdd));
-
-    try {
-      saveDocument(doc);
-    } catch (TransformerException e) {
-      e.printStackTrace();
-    }
-  }
-
-
-  private void saveDocument(Document doc) throws TransformerException {
-    TransformerFactory transformerFactory = TransformerFactory.newInstance();
-    Transformer transformer = transformerFactory.newTransformer();
-    DOMSource source = new DOMSource(doc);
-    StreamResult result = new StreamResult(resourceList);
-
-    transformer.transform(source, result);
-  }
-
-  private Element createUrlNode(Document doc, CachedFile fileToAdd) {
-    Element url = doc.createElement("url");
-    Element loc = doc.createElement("loc");
-    loc.appendChild(doc.createTextNode(fileToAdd.getFile().getPath()));
-
-    url.appendChild(loc);
-    return url;
+  public void addFile(CachedFile fileToAdd) throws ResourceSyncException {
+    ResourceSyncXmlHelper xmlHelper = new ResourceSyncXmlHelper(resourceList, this::updateMetaData);
+    xmlHelper.addUrlElement(fileToAdd.getFile());
+    xmlHelper.save();
   }
 
   private void updateMetaData(Node root, Document doc) {
@@ -90,12 +32,5 @@ class FileSystemResourceList implements ResourceList {
     } else {
       root.getFirstChild().getAttributes().getNamedItem("at").setNodeValue(resourceSyncDateFormatter.now());
     }
-  }
-
-  private Element createRootNode(Document doc) {
-    Element urlSet = doc.createElement("urlset");
-    urlSet.setAttribute("xmlns", "http://www.sitemaps.org/schemas/sitemap/0.9");
-    urlSet.setAttribute("xmlns:rs", "http://www.openarchives.org/rs/terms/");
-    return urlSet;
   }
 }
