+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/datastores/resourcesync/ResourceList.java
@@ -3,5 +3,5 @@ package nl.knaw.huygens.timbuctoo.v5.datastores.resourcesync;
 import nl.knaw.huygens.timbuctoo.v5.filestorage.dto.CachedFile;
 
 public interface ResourceList {
-  void addFile(CachedFile fileToAdd);
+  void addFile(CachedFile fileToAdd) throws ResourceSyncException;
 }
