+++ timbuctoo-instancev4/src/test/java/nl/knaw/huygens/timbuctoo/v5/dataset/DataSetFactoryTest.java
@@ -79,6 +79,14 @@ public class DataSetFactoryTest {
   }
 
   @Test
+  public void createImportManagerOnlyAddsANewDataSetToResourceSync() throws Exception {
+    dataSetFactory.createImportManager("user", "dataset");
+    dataSetFactory.createImportManager("user", "dataset");
+
+    verify(resourceSync, times(1)).addDataSet("user", "dataset");
+  }
+
+  @Test
   public void dataSetExistsReturnsFalseIfTheUserIsNotKnown() {
     boolean dataSetExists = dataSetFactory.dataSetExists("ownerId", "dataSetId");
